COMMON: Implement OrderedHashMap (WIP)
This adds a variation on HashMaps that retain the insertion
order, by keeping track of this order in an attached linked list.
This will be useful for letting the JSON-classes define the layout
of exported JSON files in terms of field order.

TODO: Some tests have been added, but a few more should
preferably be added before merging.